+++ SwiftModules/Data/globalData.swift
@@ -11,6 +11,8 @@ class globalDataModel : ObservableObject {
     @Published var selectedTab = 0  //for the main window
     @Published var spaceTotal: Double = 1
     @Published var spaceAvailable: Double = 1
+    @Published var sideBarHelpButtonText = "Explore SysMonitor's sleek interface, designed to give you quick, handy insights about your MAcBook! Right now, we're spotlighting the Drives section, where you can effortlessly check key details like available disk space and other essential info about your machine's storage. It's all about keeping you informed with essential data!" //It will hold the help info about specific section of the app, thus it is global
+    
     // @Published var spaceTotalOnExternalDrive: [String: Double] = [:]
     // @Published var spaceAvailbaleOnExternalDrive: [String: Double] = [:]