+++ SwiftModules/SysMenu.swift
@@ -15,11 +15,8 @@ struct SysMenu: View {
         
         HStack {
             
-            //custom sidebar like view
-            VStack {
-                Image(systemName: "bolt")
-                Image(systemName: "hammer")
-            }
+            sideBar()
+    
             TabView (selection: $globaldata.selectedTab){
                 
                 
@@ -72,8 +69,23 @@ struct SysMenu: View {
             }
             .tabViewStyle(.automatic)
             .padding()
+            .onChange(of: globaldata.selectedTab) { changeTheHelpMessage() }
         }
     
+    }
+    
+    
+    private func changeTheHelpMessage() {
+        
+        if (globaldata.selectedTab == 0) {
+            globaldata.sideBarHelpButtonText = "Explore SysMonitor's sleek interface, designed to give you quick, handy insights about your MAcBook! Right now, we're spotlighting the Drives section, where you can effortlessly check key details like available disk space and other essential info about your machine's storage. It's all about keeping you informed with essential data!"
+        } else if (globaldata.selectedTab == 1) {
+            
+            globaldata.sideBarHelpButtonText = ""
+            
+        }
+        
+        
     }
     
 }